Adapter: Refactor file naming
