+++ adapter/src/adapter/AdapterService.ts
